Added files via upload
I've added the refractory period; 
The way is sloppy, but works. 
I use a cell array to store the history of the active elements(because they usually have different length); 
Apparently this cell array could be used to calculate the sizes of the avalanches
Looking for a way to store only the necessary for refractory period information, but I don't have a unique solution for different refractory time yet.